Fix parser messing with input
The parser output used to be used to morph the input into something that matches our syntax. This might mess with the users ability to type some things, which is bad. Therefore, we now only store the raw text and run the parser only for verification and on proof check.

Closes #1